refator_struct
