+++ go_app/backend/main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"encoding/json" // Needed for endpoints
 	"html/template" // templating-pakke in go
 	"net/http"      // http-pakke in go
 )
@@ -21,7 +20,7 @@ func main() {
 	// TODO: Implement ^
 
 	// GET /login - Serve Login Page
-	// TODO: Implement ^
+	mux.Handle("/login", &loginHandler{})
 
 	// GET /api/search - Search
 	// TODO: Implement ^
@@ -30,30 +29,11 @@ func main() {
 	// TODO: Implement ^
 
 	// POST /api/login - Login
-	// TODO: Implement ^
+	mux.Handle("/api/login", &apiLoginHandler{})
 
 	// GET /api/logout - Logout
 	mux.Handle("/api/logout", &logoutHandler{})
 
 	// Run the server on port :8080
 	http.ListenAndServe(":8080", mux)
-}
-
-type rootHandler struct{}
-
-func (h *rootHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	// Serve Root Page
-	testTemplate.Execute(w, nil)
-}
-
-type logoutHandler struct{}
-
-func (h *logoutHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	// JSON-response
-	resp := map[string]string{
-		"status": "logged out",
-	}	
-
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(resp)
 }
